+++ packages/rum/src/domain/record/observer.ts
@@ -7,6 +7,8 @@ import {
   addEventListener,
   includes,
   DefaultPrivacyLevel,
+  isExperimentalFeatureEnabled,
+  noop,
 } from '@datadog/browser-core'
 import { NodePrivacyLevel } from '../../constants'
 import { getNodePrivacyLevel, shouldMaskNode } from './privacy'
@@ -28,12 +30,25 @@ import {
   ScrollCallback,
   StyleSheetRuleCallback,
   ViewportResizeCallback,
+  VisualViewportResizeCallback,
+  MousePosition,
+  MouseInteractionParam,
 } from './types'
-import { forEach, getWindowHeight, getWindowWidth, hookSetter, isTouchEvent } from './utils'
+import { forEach, hookSetter, isTouchEvent } from './utils'
 import { startMutationObserver, MutationController } from './mutationObserver'
+
+import {
+  getVisualViewport,
+  getWindowHeight,
+  getWindowWidth,
+  getScrollX,
+  getScrollY,
+  convertMouseEventToLayoutCoordinates,
+} from './viewports'
 
 const MOUSE_MOVE_OBSERVER_THRESHOLD = 50
 const SCROLL_OBSERVER_THRESHOLD = 100
+const VISUAL_VIEWPORT_OBSERVER_THRESHOLD = 200
 
 export function initObservers(o: ObserverParam): ListenerHandler {
   const mutationHandler = initMutationObserver(o.mutationController, o.mutationCb, o.defaultPrivacyLevel)
@@ -46,6 +61,10 @@ export function initObservers(o: ObserverParam): ListenerHandler {
   const styleSheetObserver = initStyleSheetObserver(o.styleSheetRuleCb)
   const focusHandler = initFocusObserver(o.focusCb)
 
+  const visualViewportResizeHandler = isExperimentalFeatureEnabled('visualviewport')
+    ? initVisualViewportResizeObserver(o.visualViewportResizeCb)
+    : noop
+
   return () => {
     mutationHandler()
     mousemoveHandler()
@@ -56,6 +75,7 @@ export function initObservers(o: ObserverParam): ListenerHandler {
     mediaInteractionHandler()
     styleSheetObserver()
     focusHandler()
+    visualViewportResizeHandler()
   }
 }
 
@@ -73,11 +93,16 @@ function initMoveObserver(cb: MousemoveCallBack): ListenerHandler {
       const target = event.target as Node
       if (hasSerializedNode(target)) {
         const { clientX, clientY } = isTouchEvent(event) ? event.changedTouches[0] : event
-        const position = {
+        const position: MousePosition = {
           id: getSerializedNodeId(target),
           timeOffset: 0,
           x: clientX,
           y: clientY,
+        }
+        if (isExperimentalFeatureEnabled('visualviewport') && window.visualViewport) {
+          const { visualViewportX, visualViewportY } = convertMouseEventToLayoutCoordinates(clientX, clientY)
+          position.x = visualViewportX
+          position.y = visualViewportY
         }
         cb([position], isTouchEvent(event) ? IncrementalSource.TouchMove : IncrementalSource.MouseMove)
       }
@@ -115,12 +140,18 @@ function initMouseInteractionObserver(
       return
     }
     const { clientX, clientY } = isTouchEvent(event) ? event.changedTouches[0] : event
-    cb({
+    const position: MouseInteractionParam = {
       id: getSerializedNodeId(target),
       type: eventTypeToMouseInteraction[event.type as keyof typeof eventTypeToMouseInteraction],
       x: clientX,
       y: clientY,
-    })
+    }
+    if (isExperimentalFeatureEnabled('visualviewport') && window.visualViewport) {
+      const { visualViewportX, visualViewportY } = convertMouseEventToLayoutCoordinates(clientX, clientY)
+      position.x = visualViewportX
+      position.y = visualViewportY
+    }
+    cb(position)
   }
   return addEventListeners(document, Object.keys(eventTypeToMouseInteraction) as DOM_EVENT[], handler, {
     capture: true,
@@ -141,12 +172,20 @@ function initScrollObserver(cb: ScrollCallback, defaultPrivacyLevel: DefaultPriv
       }
       const id = getSerializedNodeId(target)
       if (target === document) {
-        const scrollEl = (document.scrollingElement || document.documentElement)!
-        cb({
-          id,
-          x: scrollEl.scrollLeft,
-          y: scrollEl.scrollTop,
-        })
+        if (isExperimentalFeatureEnabled('visualviewport')) {
+          cb({
+            id,
+            x: getScrollX(),
+            y: getScrollY(),
+          })
+        } else {
+          const scrollEl = (document.scrollingElement || document.documentElement)!
+          cb({
+            id,
+            x: scrollEl.scrollLeft,
+            y: scrollEl.scrollTop,
+          })
+        }
       } else {
         cb({
           id,
@@ -337,4 +376,33 @@ function initFocusObserver(focusCb: FocusCallback): ListenerHandler {
   return addEventListeners(window, [DOM_EVENT.FOCUS, DOM_EVENT.BLUR], () => {
     focusCb({ has_focus: document.hasFocus() })
   }).stop
+}
+
+function initVisualViewportResizeObserver(cb: VisualViewportResizeCallback): ListenerHandler {
+  if (!window.visualViewport) {
+    return noop
+  }
+  const { throttled: updateDimension, cancel: cancelThrottle } = throttle(
+    monitor(() => {
+      cb(getVisualViewport())
+    }),
+    VISUAL_VIEWPORT_OBSERVER_THRESHOLD,
+    {
+      trailing: false,
+    }
+  )
+  const removeListener = addEventListeners(
+    window.visualViewport,
+    [DOM_EVENT.RESIZE, DOM_EVENT.SCROLL],
+    updateDimension,
+    {
+      capture: true,
+      passive: true,
+    }
+  ).stop
+
+  return function stop() {
+    removeListener()
+    cancelThrottle()
+  }
 }