+++ test/e2e/scenario/recorder/recorder.scenario.ts
@@ -2,10 +2,7 @@ import { CreationReason, IncrementalSource, RecordType, Segment } from '@datadog
 import { InputData, StyleSheetRuleData, NodeType } from '@datadog/browser-rum/cjs/domain/record/types'
 import { RumInitConfiguration } from '@datadog/browser-rum-core'
 import { DefaultPrivacyLevel } from '@datadog/browser-rum'
-import { renewSession } from '../lib/helpers/session'
-import { createTest, bundleSetup, html, EventRegistry } from '../lib/framework'
-import { browserExecute } from '../lib/helpers/browser'
-import { flushEvents } from '../lib/helpers/flushEvents'
+
 import {
   findElement,
   findElementWithIdAttribute,
@@ -15,7 +12,11 @@ import {
   findMeta,
   findTextContent,
   createMutationPayloadValidatorFromSegment,
-} from '../../../packages/rum/test/utils'
+} from '@datadog/browser-rum/test/utils'
+import { renewSession } from '../../lib/helpers/session'
+import { createTest, bundleSetup, html, EventRegistry } from '../../lib/framework'
+import { browserExecute } from '../../lib/helpers/browser'
+import { flushEvents } from '../../lib/helpers/flushEvents'
 
 const INTEGER_RE = /^\d+$/
 const TIMESTAMP_RE = /^\d{13}$/